Remove outdated libvirt_vif_driver settings
There is no longer a need to modify the libvirt_vif_driver conf
setting, it should always be the default value of:

libvirt_vif_driver=nova.virt.libvirt.vif.LibvirtGenericVIFDriver

This patch removes any reference to deprecated options for the
libvirt_vif_driver parameter.

Closes-bug #1192833

As a side-effect, this patch removes a section that contains this note:

      <note>
        <para>When using libvirt &lt; 0.9.11, one must also edit
            <filename>/etc/libvirt/qemu.conf</filename>, uncomment the
          entry for 'cgroup_device_acl', add the value '/dev/net/tun'
          to the list of items for the configuration entry, and then
          restart libvirtd.</para>
      </note>

It's not clear to me whether this is required or not with the
generic VIF driver plugin.

Change-Id: I233b18c8021bae6218faee63828c093d9b5b883d
backport: havana
(cherry picked from commit a31f8101cbf103916cc0b293a66ccb9bf96d46e1)