Keystone Verify - Further clarity on the use of ADMIN_PASS
A user reported an issue that the documentation was not clear as to where
the "ADMIN_PASS" paramater came from. I've clarified this further and
added a link to the relevant step to enforce it's origin. Also corrected
a very minor typo.

Note that this is now aware of the way that Debian-based platforms configure
the users in Keystone and links to the relevant page for further assistance.

Change-Id: Ie1722812386591fa0563b57a4127b0042da1ee40
backport: havana
Closes-Bug: #1252281