Added Tenant-Create step to Create Base Networks
As reported to Launchpad, the creation of the base networks
document expects the user to create one or more initial tenants.
Whilst this has been done in prior steps, it's not clear, or
obvious how to do it. I've added a basic one-liner and linked
to relevant supporting documents in the install-guide.

Changes also fix a minor typo and append the "--tenant-id"
parameter to the user input sections where required.

Change-Id: I8ef3d3c69a5b10a9a4906f4249da1605f9a01e4c
backport: havana
Closes-Bug: 1261165