Added note for Cinder Controller requirements
Although intended, the Cinder Controller document does not make it
clear as to why the 'cinder-volume' service is not installed. The
guide assumes that multiple nodes are being used; one to run all
of the controller services and many other service nodes to provide
functionality. In this case, the second part of the document
provides instructions on how to complete the process. This patch
adds a node, enhancing the first paragraph to make it clearer.

Change-Id: I0c640f5e4fca438edf214bfe00480938294c3245
backport: havana
Closes-Bug: 1260338
(cherry picked from commit 46ba7f31b6f381f877c2bea3f85d9a29170dd1c7)