Relocated step to restart Neutron plug-in agent
Relocated step to restart Neutron plug-in agent from section for
plug-in configuration to after completion of Neutron configuration.
This change allows the plug-in agent to read all configured values
from neutron.conf required for proper start and operation.

Implemented more generic phrasing with the Open vSwitch plug-in
agent as an example.

Change-Id: Ie6a614683271674678fc4d44f88c3fe47bb3b473
backport: havana
Closes-Bug: #1264163