Added step to check for /etc/neutron/plugin.ini symlink
I added a step to check for a symlink /etc/neutron/plugin.ini pointing
to the associated plug-in configuration file. Should the symlink not
exist, this step also contains instructions for creating it using Open
vSwitch as an example. This patch applies to configuring Neutron on
a controller and network node, but not a compute node since it doesn't
run the Neutron server daemon.

This patch now covers SLES and derivatives too which expect the
NEUTRON_PLUGIN_CONF variable in the /etc/sysconfig/neutron file to
reference the associated plug-in configuration file. Thanks Andreas!

Change-Id: I0ad19d479129d68543adfd65268f8925e7f0b272
backport: havana
Closes-Bug: #1260173