Readded section for Floating IP creation in admin guide
A user reported a missing section in the Havana admin guide for
creating and attaching floating IP addresses. As many users still
use nova-network I thought this should be put back in. I've added
a note to point towards the basic neutron equivalent. For the Icehouse
manual we should add more detail to support the floating IP concepts in
Neutron. Will revisit.

Change-Id: Iaa4387461a6f4775acf1f86dfde7b8afe226495f
Closes-Bug: #1263203