Fix systemctl problem on SLES in ceilometer setup
There is a problem when enable mongodb service on SLES in ceilometer
install section, which uses `systemctl start mongodb.service`.

systemctl is not installed in SLES SP3 by default, and
mongodb.service is not a valid service too. This patch uses common
commands of `service` and `chkconfig` to do the right job.

Change-Id: I3d44f6aba155601030b9b13dc7ef29550e19dc2c
Closes-Bug: #1267297