Clarified database configuration steps for Block Storage
I clarified the database configuration steps for Block Storage for
consistency with the associated patch applied to Havana. I also added
a note for Ubuntu users to add the '[database]' section header if not
included with the default "cinder.conf" file. This patch covers both
the controller and storage nodes.

Change-Id: I6ae7072b2a98db00a6367213721f883072f8348b