Removed package providing 'neutron-server' from network node
While fixing another bug, I noticed that the instructions for RHEL and
SUSE install a metapackage providing 'neutron-server', but do not
configure it to start on boot. In comparison, the Ubuntu instructions
install the specific package providing 'neutron-server' and
automatically configure it to start on boot.

I originally intended to patch this issue by providing instructions
for disabling 'neutron-server' in Upstart, but upon further
investigation, determined that 'neutron-server' does not need to run
on the network node, nor do any of the other packages
(neutron-l3-agent and neutron-dhcp-agent) depend on the neutron-server
package. So, this patch removes the neutron-server package from the
Ubuntu instructions and from the list of services to restart on all
distributions. I also removed the neutron-plugin-openvswitch-agent
package from the Ubuntu instructions since we shouldn't cover specific
plug-ins in this section.

Change-Id: I5c08a56b37d8aef656826fc1b644efc799aa51a7
backport: havana
Closes-Bug: #1265833