Removed packages installed as dependencies by OVS plug-in agent
I removed packages installed as dependencies by the Open vSwitch
plug-in agent package on Ubuntu and Debian. For example, the
neutron-plugin-openvswitch-agent package automatically installs
the openvswitch-switch package, so the instructions shouldn't
specifically call out installing the latter.

Also, on the controller node which doesn't need to run the OVS
plug-in agent but still needs to know about the configured plug-in,
I changed the neutron-plugin-openvswitch-agent package to the
neutron-plugin-openvswitch package.

Change-Id: I13a50854149109a10155d0b6b05307a663563438
backport: havana
Closes-Bug: #1264184