Added configuration for Glance to authenticate via Keystone
Apparently the existing instructions lack the configuration keys
necessary for Glance to authenticate via Keystone which breaks various
features. I added 'flavor = keystone' to glance-api.conf and
glance-registry.conf to resolve this issue. However, this generated
a warning about the lack of 'auth_uri' which I also added to resolve
this issue. This patch also includes general clarifications.

Since this patch plays a significant role in building a functional
environment, I recommend considering it for backporting to Havana.

Change-Id: If112a7d039f3943238e915ef0066765529c7d668
backport: havana
Closes-Bug: #1262755
Closes-Bug: #1267854