Added step to restart Nova after configuring it for Neutron
I added a step to restart the Nova API service on the controller node
after configuring Nova to use Neutron for networking. We should
consider backporting this patch to Havana since the issue probably
leaves the controller node in an indeterminate state until the
operator restarts services or reboots the system during the
troubleshooting process.

After some discussion with Steve Gordon, I updated this patch to also
include restarting nova-scheduler and nova-conductor for completeness.

Change-Id: I63c25aa285616ee95ecde2d685f52103809adec7
Closes-Bug: #1269253