Edits, fix bullet list
As mentioned in the bug, the text "As a user" should not be in the list
but is the introduction to the next list.

Patch fixes also two whitespace problems, the first noticed by the
niceness gate, second by visual inspection.

Patch also fixes spelling errors and a few wording issues. (diane fleming)

Co-Authored-By: Diane Fleming <diane.fleming@rackspace.com>

Change-Id: I687b23cf19789c1039029984133cb7b005521374
Closes-Bug: #1269681