Replace the_service_id_above with shell script
To create an endpoint, user should copy the command to text editor,
then copy the service id and replace the literal named
the_service_id_above with the real id, then copy the whole command to
shell and execute it.

It is not so convenient, so I add a short shell script to reduce this
labor work.

Change-Id: I4992b76e574536224896562b3599e514881c9f76
Closes-Bug: #1271637