Improve publishing of Admin Guide Cloud
Before this change, we have been publishing the whole of webhelp
to the website which is:
du -hs target/docbkx/webhelp/*
11M	target/docbkx/webhelp/admin-guide-cloud
2.3M	target/docbkx/webhelp/autopdf
836K	target/docbkx/webhelp/bk-admin-guide-cloud.xml
6.8M	target/docbkx/webhelp/cloud
22M	target/docbkx/webhelp/fonts
1.9M	target/docbkx/webhelp/images

This patch now cleanes up the publishing so that webhelp only
contains the admin-guide-cloud directory as it should.

The patch also does some minor cleanup of the pom file and removes
an unneeded generate-pdf execution.

Note: If this patch gets accepted, I will look at the other pom files
as well.

Change-Id: Iafcb227d9a65776f0beb9750463f7b54e9f05269