Start the nova-network service instead of restarting it (minor edit)
The fourth (4) step of 'Configure networking' in the 'Configure Compute services'
chapter (RHEL, CentOS, Fedora), says: start the network service, but in the command
of the step it restarts the network service. This results to a warning that the
openstack-nova-network service was not able to stop.
This commit just changes the argument of the service command to start.

Change-Id: I348f63d89740f9935390e4861b68fe48f8fb24cf
backport: havana