vmware: Updates in data store selection section
The VMDK driver was selecting datastores for volume creation based
on space utilization. After the fix for bug 1250816, the datastore
which is connected to maximum number of hosts is selected for volume
creation. Ties are broken based on space utilization; datastore with
low space utilization is preferred. This fix updates the data store
selection section in the VMDK driver configuration to reflect this
selection logic.

Closes-Bug: #1253595
Change-Id: Ia562c452e8cf6bb1c8be6afd868f4b0b2ce0b006