Adds clarity to the usage of the authorisation token
As reported in LP#1277289, the usage of the authorisation
token is not clear in the instructions. It refers to a token
previously created but potentially leaves the user
confused as to which one to use. This patch adds the clarity
required and posts a link to the documentation section.

Change-Id: Ib16bd2770d4c4aac5292e9a3d2f25c2417e818f2
Closes-Bug: #1277289
backport: havana
(cherry picked from commit c2e6c077f31d29bd21e743a2ded36916857b14bb)