Fixes root_helper copy and paste problem with openstack-config
The documentation asks the user to use openstack-config to apply
a change to the root_helper parameter. With the way that
openstack-config works it expects parameters to be organised properly.
Therefore, any white space between arguments is assumed to be
a seperate argument, hence it is ignored; the solution is to
parenthesise the parameter required, encompassing all white space.

Note: Rebased since https://review.openstack.org/#/c/74243/

Change-Id: I2ab702d97828311f3fa409b6cb88931cd5f6c315
Closes-Bug: #1281565
backport: havana