Remove ovs_use_veth from cloud-admin-guide
Remove direction to set ovs_use_veth to true from cloud-admin-guide.
It's not required for recent kernels, and for those who aren't using
recent kernels we catch it in the install guide.

As this guide came into being after all supporting distributions had
moved to provide namespaces support I do not think we need to cater to
that situation here.

Change-Id: Ie0fa23bfdfcc3454988ead175ef0388f6a0602ba