Add docs for NetApp copy offload tool in Cinder
This change documents the addition of the NetApp copy offload tool into
the NetApp unified driver within Cinder. An additional Cinder
configuration option was added and its use in conjunction with the copy
offload binary is discussed, along with instructions to download and
configure the storage system to support the copy offload workflow.

Change-Id: Ie3562a83c0f9c6ff4a786eb899fdbf1e5c6efcf3
Implements: blueprint copyoffload