Updated 'Launch instances' section in the End User Guide
Included the new options available in the Launch an Instance window for
launching an instance in the Dashboard chapter of End User Guide. Also removed
link to the 'launch instance from volume' section as the steps seems redundant.

Add a .htacces redirection to handle the removed file (boot from
volume).

Partial-Bug: #1279133

Change-Id: I7f29aa45211d84e2dfab772235891a776492686d