Add docs for new HP LeftHand Cinder driver
This change adds the documentation for the new HP LeftHand
driver. It documents how to configure the driver in both
standard and legacy modes, and describes the functionality
when running in each mode. When the driver executes in legacy
mode, it is essentially running the old driver code.
Therefore, the old LeftHand driver documentation was migrated
into the section “HP LeftHand/StoreVirtual CLIQ driver legacy mode”.

Closes-Bug:#1279030
Implements blueprint lefthand-cinder-driver

Change-Id: I39d9d48669b9b7192814a5e14f0f81a8557d1a44