Clarify VMware ephemeral disks expectations.
The previous text indicated that ephemeral disk support was expected in
an upcoming stable release and the lack of such support was temporary in
nature.

As this looks to have been the case for the life of this driver, and
the fix recently pushed out to Juno, I'm updating to clarify
expectations.

In particular:

* The fix is expected in an upcoming major release, not a stable branch
  update.
* Given the above the situation is not temporary.

Change-Id: I1910dc00104fc3eaf26055623c4611c42bdb5338