Update flags and tables for neutron
Remove unneeded xi:include (metadata agent is not a plug-in)

Provide tables for new plugins and ML2 mechanisms

To avoid breaking the large tables appearance by using xrefs, the
bigswitch table is generated twice (one table for the ML2 mechanism, one
table for the standalone plugin). This is a (hopefully) temporary
change. Tables generations will probably be updated in Juno.

Change-Id: I27573975c83c5b25682135a4ea3ab852fec170ea
Closes-Bug: #1291785
Closes-Bug: #1292317
Closes-Bug: #1292318
Closes-Bug: #1294402
Closes-Bug: #1294405
Closes-Bug: #1294961