fix erroneous command in cloud admin guide
I believe that the initial "iptables -L -nv" command in this section is
erroneous and should simply be dropped.

The task is:

  Ensure the NAT rules have been added to iptables on the node that
  nova-network is running on, as root

NAT rules only exist in the iptables nat table.  There are two commands
following this task; the first one (removed by this patch) refers to the
*filter* table and shows impossible output (the filter table cannot
contain a DNAT rule).

The second iptables command correctly specifies "-t nat" and appears to
contain the same information as the impossible output from the previous
command.

Change-Id: I8dce50b6e1a1e42ff3955ecb3fbc612cd4a9f074
Closes-bug: 1294196