Flatten post-install steps for swift install
Post-install steps section is an old label from the rpevious install
guide layout. This change bring swifts in-line with other parts of
the guide, which use the new layout.

Existing content labels do not change, meaning they are still
searchable and linkable as before, but an htaccess modification
was added for the index page.

Change-Id: I8ac3f93b70928f5ce07203ff30a5d59fb927d72e