Add a note that the glance-registry is internal
Users could be confused into thinking the glance registry
is an external-facing service. It is not, and is designed
with a security model such that it should be protected for
internal use only.

This patch adds a note to the introduction in the common section
so it will be included in multiple guides.

Change-Id: Ic540353d82c829475ac6f3455ccccdea32977a4b
Closes-Bug: 1252931