Provide rootwrap configuration options
Our tools now generate tables for the rootwrap.conf files. Include this
tables in the config-reference guide. The cinder-rootwrap.xml file is
not used ATM, because our cinder documentation doesn't provide a config
overview (see bug #1304172).

Revert some help strings for nova to what they are in icehouse (the
previous run of the tool accidentally picked up the juno options).

Change-Id: Id3a76cc5ae29b7cf45118958520126ec8f75853d