Add info on how to add Database service (Trove)
The install guide did not have a chapter on how to add the
Database service. Created one with the following sections:
overview, install, verify.

This draft reflects the fact that some technical issues are still
being worked out:

-- image creation questions. Before you can use Trove, you need to
create an image for each flavor of database you want to have
(MySQL, MongoDB, etc.) The appropriate level of documentation
for this step is still under discussion and has not yet been
determined.

Change-Id: I31e02f48975c7050022ee6fb8f31b878c6b508ff
Partial-Bug: #1304584