Updated nova networking sections
According to bug #1284633, the configuration included with nova
packages on SUSE apparently doesn't default to nova-networking. We
should avoid implicit configuration assumptions on all distributions
and explicitly configure the following keys in nova.conf on the
controller and compute nodes for nova-networking:

network_api_class = nova.network.api.API
security_group_api = nova

Prior to this issue, we only included compute node configuration for
nova-networking. I added a new section for the controller node to
resolve this issue.

This patch also includes some minor structural changes and
clarifications to conform with other portions of the guide.

Change-Id: I8224f32062f279b6032aa06ee0a72ea28f549a4d
Closes-Bug: #1284633