Run db_sync commands as the proper users
keystone-manage db_sync creates a log file
/var/log/keystone/keystone.log (and other db_sync commands create similar
logs in their respective directories).  If run as root, the log file is
owned by root.  In some cases (keystone), this prevents services from
starting.  In other cases (e.g. glance-api), it prevents the service
from writing to its log file.  Thus, all db_sync commands were changed
to run them as the correct users for the services being configured.
Note: this does not appear to be a problem with Ubuntu, but I made the
change for all distros for consistency, since it will not harm anything
to run db_sync as the correct user on Ubuntu, even if it is not
necessary.  Feel free to -1 this if you think there should be an
OS-specific step instead.

Change-Id: Idd8ea62cf883edcc7b308d25292d8b18d2631330
Closes-Bug: #1308316