Fix indistinct text in cinder control install
User complains that the database configuration text for cinder control
in install guide is indistinct. He/She was confused by the words 'the
following key' and there is no such key mentioned in this step. This may
be caused by we didn't update the text when we improve this step with
openstack-config command.

Change-Id: I57c901eef2fcf694a897d117bfe662534bd27e63
Closes-Bug: #1309755
backport: havana