Use keystone instead of nova as example
If this section is used in the Install Guide, keystone is installed
but nova not yet. Thus change the example to use keystone.

Also adjust parameter to use --os-password instead of --password and
use "service-list" as keystone command. Explain that this is not
specific to keystone or nova but to all client commands.

Change-Id: I308ba102d1217acd0a9f723643728a3e2af7f77c
Closes-Bug: #1310375