Remove gateway definitions in interfaces
The interfaces templates for /etc/network/interfaces contain 2 gateway
definitions, and there is a third one configured via DHCP when the
network comes up.

Having multiple default gateways (at least configured this way) does
not work. For instance, booting into Ubuntu 12.04.4 LTS with one of
these interface definitions results in error messages ("Waiting up to
60 more seconds for network configuration", "Booting system without
full network configuration").

With the gateways removed, the network comes up without a problem and
the hosts can communicate over the API Network, the OpenStack Management
Network, and the Data Network. (And there is still a default gateway
on the NAT interfaces of each host).

Change-Id: I3ee9737c19b696d88f9897547996cddef482b7f7