Keep Glance.sh from using the Internet
The scripts running after PreInstall.sh are supposed to work without
an Internet connection ("Internet connection is not required for this
script to run").

However, Glance.sh tried to wget the CirrOS disk image which cannot
work in an offline environment. This patch moves the download into
PreInstall.sh. Also, update CirrOS image from 0.3.1 to 0.3.2.

Change-Id: Ibefbffdda1591ed4561a67edda8a50f17c14aae5