Replace X with check marks
This changeset replaces the character "X" with a check mark for the
table "Additional Security Features" in chapter "40. Hypervisor
Selection" of the security-guide.

It is inspired by ch_arch_storage.xml in the OpenStack operations-guide,
but puts the entity declaration into a separate, new file in
doc/common/entities where it can be referenced as a parameter entity
in all manuals that need it.

Change-Id: Iea8d82163d11eb2cf5b7e4f086a530de2c06f350