Cleanup section_nova-conf.xml
With the addition of section_config_format, we can now remove
section_nova-conf.xml. It was the only section in the guide that
documented the INI format.

Move some content from section_nova-conf.xml to
section_config_format.xml and update both as appropriate.

Also, update section_config_format.xml to use spaces around "=" and to
not chunk the sections.

Change-Id: I41f6e919bf996cb28a6f1ca4425752a653a3f81f