Added missing important step to Glance config
The section covering the use of vCenter datastores for the Glance
backend was missing a key step about creating a storage profile.
This section has been updated to include this step. I also added
the stop chunk command to the vmdk driver doc to keep the sections
together.

Change-Id: I68d55ba6948d1083032a8ee347b81d2b4f63e980
backport: icehouse
Closes-Bug: #1320369