Do not use root user to flush Keystone tokens
If you run keystone-manage token_flush as root and you have verbosity
enabled, the log file keystone.log will be owned by root user if it does
not exist. If Keystone is not started, it won't be able to start the
first time due to permission issue.

Using "keytone" user by default to flush the logs avoids to face this
situation and let the user know he has to use OpenStack users when
running CLI.

Change-Id: I9c806d13c727a79869485cd5f5d5d51b757a783e
Signed-off-by: Emilien Macchi <emilien.macchi@enovance.com>