Convert HA Guide to XML
This patch does the initial conversion of the High Availability
Guide to XML from ASCIIdoc.  No changes to the structure have
been made; this is simply the XML that gets generated by asciidoc.

This patch includes the following changes to the original HA
Guide:

- Content now appears in a single Docbook XML file, to be
  broken out in the next patch
- Original ASCIIDoc files have been removed
- pom.xml has been updated to reflect the new source
- .gitignore has been updated so that it no longer ignores
  bk-ha-guide.xml

This patch is simply to establish the new structure.  To prevent
front-end disruptions it includes all of the existing content,
but the content hasn't been audited or broken out into individual
files.  For ease of review, that will be done in the next
patch(es).

(This patch also includes the corrections and changes that were
originally part of https://review.openstack.org/#/c/93856/2,
corrected so that list appears properly.)

Implements: blueprint convert-ha-guide-to-docbook

Change-Id: I0455716c037b5e6e698bfca72162aba29f339e8a