Fix neutron ml2 procedure
In Debian, debconf takes care of selecting the correct plug-in, and
there is no specific plug-in packages (they are all contained within
python-neutron and neutron-common). This patch fixes the procedure
for Debian, and also displays the screenshot which was removed
somehow.

Note: this is a backport of https://review.openstack.org/#/c/95454/
without a single modification.

Change-Id: Ic50d40951e37ee41c118d980ae648f12906d8add