Documents what debconf does
After discussing the topic on the doc list, it appears that it
would be nice to document what the debconf screens are configuring.
This patch adds the information, so that it makes it more explicit
to our users what Debian does for them.

This patch only addresses the debconf chapter. There will be other
patches for specific services later on, once this patch is merged.

Change-Id: I60c7ba818e36056ed913582e66bf24c41b2c6807