neutron: admin guide cloud: updating an obsolete note on allowed address pairs
There is an obsolete note in "admin-guide-cloud" in networking introduction section
regarding neutron not allowing allowed-address-pair input that overlaps with a
fixed-ip on the same port. A recent patch to Neutron
(https://review.openstack.org/#/c/94508/2) removes this restriction.
This guide updates the note about allowed-address-pair to indicate to which
releases it applies.
Also, clean up formatting of procedures and formal para's.
Also, clean up intro text to steps and procedures.
Also, fix other odd formatting.

Change-Id: Ic89bee754a05fdb26b966b5714c729c4bc05bb50
Closes-Bug: #1327359