Make Juno-Trunk relationship clearer
Recently we've had some people trying to install Icehouse with
the Trunk guide. This no longer works.

In this patch, changes are made which will deliberately break the guide
by changing the package repositories to Juno versions that do not exist
yet.

The hope is that this will encourage people to seek out the most appropriate
guide for them - likely icehouse.

Change-Id: I42393f466c910e1d2cb350134efa815184c66902