Add steps for restarting Message Bus on Fedora
The messagebus service is renamed dbus in Fedora. As such the
instructions have been updated to provide both the RHEL/CentOS
and Fedora variants.

Examples of the service/chkconfig commands were not switched to
systemctl in this change as service/chkconfig remains used
consistently across the guide at this time and works on both
RHEL/CentOS-based and Fedora distributions.

Closes-Bug: #1322887
backport: icehouse
Change-Id: I9a7dc7fcbb11b7e342aaeebe4d57b95709a92363