Enable building of French user Guide
The user-guide has been mainly translated to French with now 76 per cent,
add it to build jobs.

For building, we need the glossary as well, thus import it manually (3
per cent translated, thus will not be downloaded automatically) and
update the Japanese glossary as well.

Import manually the current version of doc/user-guide/locale/fr.po since
it contains fixes to get this building.

Change-Id: I11280778bbf53a21e008d0dcfe4d21bfc89d7aa1