Add alternative to multiple floating ip pools supported by commit 90250
With the addition of onlink routes for all subnets on an external network to a router,
it is now possible to support multiple floating ips without having to create multiple
external networks.  Added documentation section describing this alternative and
providing example code for adding multiple subnets to an external network.

Rather than immediately deprecate the existing method, moved that to be an alternative
for future deprecation.

Made some edits for consistency

Change-Id: I50ca6fb81eb2499f2f334554c454ff1fc57f5476
Closes-Bug: #1323984