Changed order of arguments
The order of arguments for the following trove commands
was changed:

trove backup-create

trove configuration-attach

This change is documented in the following bug:

https://bugs.launchpad.net/openstack-manuals/+bug/1325121

What changed is that now (for both cmds), you pass in the
instance ID as the first arg.

The following doc page contains examples that show the
old syntax:

http://docs.openstack.org/user-guide/content/trove-manage-db.html

I updated the doc so it shows the new syntax.

Change-Id: Ic7fd24947332500ce102c3902a7ec84be4f3d542
Partial-Bug: #1325121