Re-word the Building Blocks first paragraph
The first paragraph of the building block section of the compute admin
guide implies that an ephemeral instance will loose it's state on
shutdown of the instance.  Shutdown may imply deletion, but as there
is a shutdown state that does not remove the ephemeral volume, it
is more appropriate to talk about the deletion of the virtual
instance.  This change addresses that inconsistency.

Removed excess spaces
Removed Nova-Volume reference

Change-Id: If00ef49fecdf864f74efc7171850bbacd60dbcd0