Update glossary meta data
For syncing the glossary between repositories, let's use the same xml:id
everywhere. Let's use openstack_glossary here as this is the name used
in the OPS guide.

Also, use the content about updating the glossary that is used in the
OPS guide and improve it a bit.

implements bp common-glossary-setup

Change-Id: Icf070873b49c28ba3703e9443b69ae6515b63a1f