Setup for syncing of glossary
These files are used by an infra script to sync files from
openstack-manuals to security-doc and operations-guide repositories.

The file projects-to-sync.txt contains a list of repositories to update
and sync-projects.sh handles the copying of files. This is a very simple
script to get the job done.

implements bp common-glossary-setup

Change-Id: I55be62393213a5ea2657face0f6f5f8bed1b982c