Clean Common
This patch works several files that were not in common use back
into the documents they should live with.

The intended end result of this activity is to make common as lean as
possible so it can be more easily translated.

Also: Don't include configdrive table in user-guide since it's not
relevant for audience

Change-Id: Iacd5c68c639e60e8c711cb18351543e477444299