Arch Design: Add glossentries, use common glossary
Use the common documentation glossary.
Move entries from ch_glossary to the common documentation glossary and
add glossterm so that the entry gets displayed in the glossary.
Remove entries from the Arch Design specific glossary that are fully
explained in the text.
Remove entries that are already in the common glossary and add glossterm
for them in the text.
Add glossary as final chapter to the manual.

The remaining glossary entries need some further discussion.

Change-Id: Ic82ae5235e1306fc13be81fa4971710321da5072