Removes unnecessary changes to OVS agent config file
In the existing documentation it suggests that the OVS agent needs
to use the ML2 configuration file, however this is not the case. The
Neutron Server needs to use the ML2 configuration, but the OVS agent
local to the Neutron node or Compute node needs direct access to the
OVS configuration file.

The existing documentation guides the user to run a sed replacement
of the initscript (or systemd service) for it to work succcessfully
and the raw code points to a Red Hat bugzilla entry.

This shouldn't be backported, but another patch for stable/icehouse
should be implemented.

Change-Id: I334127add2816c1967830f58f11fd806cebc71c7
Closes-Bug: #1356290