Adds note for unexplained error messages
In the stable/icehouse branch for RHEL-based deployments,
it asked the user to start and enable dbus/messagebus. In
RHEL 7 this is not required as it's out of the box.

As Icehouse standardised on RHEL 6, we should keep it as
it is, but it's confusing if someone's deploying Icehouse
on RHEL 7 based platforms. This patch adds a note to advise.

References https://review.openstack.org/#/c/115753/

Change-Id: Ic16ff621b309aca270bca8bfe833553889a21c3d