Modifies example command for nova network-create
The example given for creating networks in the Networking
concepts page in the OpenStack Cloud Administrator Guide
has an error. The --fixed-cidr should be a subset of the
--fixed-range-v4, but the example fixed-cidr is both larger
and part of a different network.

This fix corrects the subnet sizes and changes the network
addresses on both --fixed-range-v4 and --fixed-cidr to make
--fixed-cidr a valid subnet of --fixed-range-v4.

Change-Id: Ibb933bea845dee6cd1a5f0559e392820ff376a84
Closes-Bug: #1366214