Python 2.6 is end-of-life
Python 2.6 is end-of-life, and we should be recommending Python 2.7.

This patch updates two of the three places Python 2.6 was referred to
in the documentation, removing the reference to 2.6 but keeping the
2.7 reference. The final place is in the debian packages section of
the install guide, which notes several issues with Python 2.7 that
could now be solved. If this is the case, this entire section should
probably be removed.

Change-Id: I8c4043d180bc47de2afc40ce73640982f4c2f663