Deprecate / obsolete NetApp volume extra specs
The NetApp Data ONTAP (Cluster-mode) NFS & iSCSI drivers for Juno support
the Cinder pools feature, but the drivers are reporting two qualified
extra specs that must be converted to unqualified extra specs in order to
be used by the Cinder scheduler's capability filter. Furthermore, there
are four extra specs that must be deprecated due to having the pools
feature.

This patch updates the documentation to update the extra specs and
formally denote the deprecation.

Change-Id: I36ab7ed86bdad09ff25c6f2cc280212ca8d52b98