Use UTF-8 as charset for HTML pages
According to the W3C validator UTF-8 should be used instead of utf8.

---snip---
Rare or unregistered character encoding detected

The character encoding declared for this document (utf8) is supported by the
validator, but may not be widely supported across platforms. For the sake of
interoperability, it is best to use a unicode character encoding such as UTF-8,
or one of the registered character encodings.
---snap---

Change-Id: Ife1862c7c245ee610a7a4d820f9717626e56244c