Rebuild the hot-ref with new IDs
This patch uses the content generated with a new version of dn2osdbk.
All the docbook IDs have been changed, and are not prefixed with the
source file name. This should not be a problem since this guide has not
been officially released and is not linked from the docs index page.

Change-Id: Ie781bd340428bf1f56f7c4dcb80ea506de4ba8f3