Added doc for configuring clusters
An admin user can modify various cluster
characteristics by editing the
/etc/trove/trove.conf file.

This commit covers how to set the number of
query routers and the number of config servers
per cluster.

This commit is a partial fix. Once the following
gets merged:

https://review.openstack.org/#/c/120672

the doc can be expanded to cover how to set
the flavor IDs for config servers and
query routers.

Change-Id: Ibc90e449764143b1145d4fc1d536f40f13c46ac3
Partial-Bug: #1392419