Update glance-api sample
as noted in the bug report, notifier_strategy was removed.
The change was picked up by the auto config tools, but the
 remaining place this was used in manuals was the glance-api
install guide samples. This patch updates the samples from
the glance master, and makes the small number of changes
(flavor, keystone_authtoken setup) that were in the previous
samples.

Change-Id: I673590ed2470d07cce93ed5713c7bffe7f57730c
Closes-Bug: 1332301