Ignore <option> tags in generatepot
With https://review.openstack.org/#/c/144685 merged,
the automatically generated configuration reference
tables will be marked up as
<option>option_name</option> = <replaceable>default_value</replaceable>

As none of this should be translated (option_name is the same
regardless of language), this patch updates generatepot to ignore
option_name s tagged with <option> for translation.

This should significantly reduce the number of strings required for
translation from the common directory in particular, where
many of the 8000-odd strings do not actually need translation.

Change-Id: I7b46f4dd10c0855f86cea35417dcfde93cc113a3