Add info on what does the debconf system
As per the discussion in the meeting today, it seems needed to let
users know what the debconf system will do when installing Keystone.
This patch fixes it.

This patch also re-adds the "Create tenants, users and roles" chapter
to the Debian install-guide, and adds a note explaining that the
admin user, tenant and roles can be created automatically by the
Keystone package.

Also, Debian configures a /etc/cron.hourly/keystone, so the part which
describes how to manually configure it has been fixed for Debian.

Change-Id: I2e828b619ba9c8c1a55c2a4ea8de8bae691d5f2d