Provide References for Cinder Quota Defaults
The documentation currently does not reference what the cinder quota
default values are or where they are found.

Adds documentation covering which file the values are found in,
where in the file they are found, and a link to a sample cinder config
file so users can quickly find which values exist in the default config.

Change-Id: Ie8ebe52764951b5907f0c40997251176fa8e023a
Closes-Bug: #1374156